+++ src/main/java/persistence/entity/DefaultEntityManager.java
@@ -1,5 +1,6 @@
 package persistence.entity;
 
+import jdbc.DefaultRowMapper;
 import jdbc.JdbcTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
