+++ src/test/java/jdbc/InstanceFactoryTest.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package jdbc;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
